DS4A-final-be/simple-deploy/requirements.txt
Flask
Pillow
Numpy
tensorflow
